+++ azure-cognitiveservices-vision-face/azure/cognitiveservices/vision/face/operations/face_operations.py
@@ -35,8 +35,31 @@ class FaceOperations(object):
 
     def find_similar(
             self, face_id, face_list_id=None, large_face_list_id=None, face_ids=None, max_num_of_candidates_returned=20, mode="matchPerson", custom_headers=None, raw=False, **operation_config):
-        """Given query face's faceId, find the similar-looking faces from a faceId
-        array, a face list or a large face list.
+        """Given query face's faceId, to search the similar-looking faces from a
+        faceId array, a face list or a large face list. faceId array contains
+        the faces created by [Face -
+        Detect](/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f30395236),
+        which will expire 24 hours after creation. A "faceListId" is created by
+        [FaceList -
+        Create](/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f3039524b)
+        containing persistedFaceIds that will not expire. And a
+        "largeFaceListId" is created by [LargeFaceList -
+        Create](/docs/services/563879b61984550e40cbbe8d/operations/5a157b68d2de3616c086f2cc)
+        containing persistedFaceIds that will also not expire. Depending on the
+        input the returned similar faces list contains faceIds or
+        persistedFaceIds ranked by similarity.
+        <br/>Find similar has two working modes, "matchPerson" and "matchFace".
+        "matchPerson" is the default mode that it tries to find faces of the
+        same person as possible by using internal same-person thresholds. It is
+        useful to find a known person's other photos. Note that an empty list
+        will be returned if no faces pass the internal thresholds. "matchFace"
+        mode ignores same-person thresholds and returns ranked similar faces
+        anyway, even the similarity is low. It can be used in the cases like
+        searching celebrity-looking faces.
+        <br/>The 'recognitionModel' associated with the query face's faceId
+        should be the same as the 'recognitionModel' used by the target faceId
+        array, face list or large face list.
+        .
 
         :param face_id: FaceId of the query face. User needs to call Face -
          Detect first to get a valid faceId. Note that this faceId is not
@@ -121,7 +144,22 @@ class FaceOperations(object):
 
     def group(
             self, face_ids, custom_headers=None, raw=False, **operation_config):
-        """Divide candidate faces into groups based on face similarity.
+        """Divide candidate faces into groups based on face similarity.<br />
+        * The output is one or more disjointed face groups and a messyGroup. A
+        face group contains faces that have similar looking, often of the same
+        person. Face groups are ranked by group size, i.e. number of faces.
+        Notice that faces belonging to a same person might be split into
+        several groups in the result.
+        * MessyGroup is a special face group containing faces that cannot find
+        any similar counterpart face from original faces. The messyGroup will
+        not appear in the result if all faces found their counterparts.
+        * Group API needs at least 2 candidate faces and 1000 at most. We
+        suggest to try [Face -
+        Verify](/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f3039523a)
+        when you only have 2 candidate faces.
+        * The 'recognitionModel' associated with the query faces' faceIds
+        should be the same.
+        .
 
         :param face_ids: Array of candidate faceId created by Face - Detect.
          The maximum is 1000 faces
@@ -182,6 +220,36 @@ class FaceOperations(object):
             self, face_ids, person_group_id=None, large_person_group_id=None, max_num_of_candidates_returned=1, confidence_threshold=None, custom_headers=None, raw=False, **operation_config):
         """1-to-many identification to find the closest matches of the specific
         query person face from a person group or large person group.
+        <br/> For each face in the faceIds array, Face Identify will compute
+        similarities between the query face and all the faces in the person
+        group (given by personGroupId) or large person group (given by
+        largePersonGroupId), and return candidate person(s) for that face
+        ranked by similarity confidence. The person group/large person group
+        should be trained to make it ready for identification. See more in
+        [PersonGroup -
+        Train](/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f30395249)
+        and [LargePersonGroup -
+        Train](/docs/services/563879b61984550e40cbbe8d/operations/599ae2d16ac60f11b48b5aa4).
+        <br/>
+        Remarks:<br />
+        * The algorithm allows more than one face to be identified
+        independently at the same request, but no more than 10 faces.
+        * Each person in the person group/large person group could have more
+        than one face, but no more than 248 faces.
+        * Higher face image quality means better identification precision.
+        Please consider high-quality faces: frontal, clear, and face size is
+        200x200 pixels (100 pixels between eyes) or bigger.
+        * Number of candidates returned is restricted by
+        maxNumOfCandidatesReturned and confidenceThreshold. If no person is
+        identified, the returned candidates will be an empty array.
+        * Try [Face - Find
+        Similar](/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f30395237)
+        when you need to find similar faces from a face list/large face list
+        instead of a person group/large person group.
+        * The 'recognitionModel' associated with the query faces' faceIds
+        should be the same as the 'recognitionModel' used by the target person
+        group or large person group.
+        .
 
         :param face_ids: Array of query faces faceIds, created by the Face -
          Detect. Each of the faces are identified independently. The valid
@@ -260,6 +328,17 @@ class FaceOperations(object):
             self, face_id1, face_id2, custom_headers=None, raw=False, **operation_config):
         """Verify whether two faces belong to a same person or whether one face
         belongs to a person.
+        <br/>
+        Remarks:<br />
+        * Higher face image quality means better identification precision.
+        Please consider high-quality faces: frontal, clear, and face size is
+        200x200 pixels (100 pixels between eyes) or bigger.
+        * For the scenarios that are sensitive to accuracy please make your own
+        judgment.
+        * The 'recognitionModel' associated with the query faces' faceIds
+        should be the same as the 'recognitionModel' used by the target face,
+        person group or large person group.
+        .
 
         :param face_id1: FaceId of the first face, comes from Face - Detect
         :type face_id1: str
@@ -318,9 +397,49 @@ class FaceOperations(object):
     verify_face_to_face.metadata = {'url': '/verify'}
 
     def detect_with_url(
-            self, url, return_face_id=True, return_face_landmarks=False, return_face_attributes=None, custom_headers=None, raw=False, **operation_config):
-        """Detect human faces in an image and returns face locations, and
-        optionally with faceIds, landmarks, and attributes.
+            self, url, return_face_id=True, return_face_landmarks=False, return_face_attributes=None, recognition_model="recognition_01", return_recognition_model=False, custom_headers=None, raw=False, **operation_config):
+        """Detect human faces in an image, return face rectangles, and optionally
+        with faceIds, landmarks, and attributes.<br />
+        * Optional parameters including faceId, landmarks, and attributes.
+        Attributes include age, gender, headPose, smile, facialHair, glasses,
+        emotion, hair, makeup, occlusion, accessories, blur, exposure and
+        noise.
+        * The extracted face feature, instead of the actual image, will be
+        stored on server. The faceId is an identifier of the face feature and
+        will be used in [Face -
+        Identify](/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f30395239),
+        [Face -
+        Verify](/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f3039523a),
+        and [Face - Find
+        Similar](/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f30395237).
+        It will expire 24 hours after the detection call.
+        * Higher face image quality means better detection and recognition
+        precision. Please consider high-quality faces: frontal, clear, and face
+        size is 200x200 pixels (100 pixels between eyes) or bigger.
+        * JPEG, PNG, GIF (the first frame), and BMP format are supported. The
+        allowed image file size is from 1KB to 6MB.
+        * Faces are detectable when its size is 36x36 to 4096x4096 pixels. If
+        need to detect very small but clear faces, please try to enlarge the
+        input image.
+        * Up to 64 faces can be returned for an image. Faces are ranked by face
+        rectangle size from large to small.
+        * Face detector prefer frontal and near-frontal faces. There are cases
+        that faces may not be detected, e.g. exceptionally large face angles
+        (head-pose) or being occluded, or wrong image orientation.
+        * Attributes (age, gender, headPose, smile, facialHair, glasses,
+        emotion, hair, makeup, occlusion, accessories, blur, exposure and
+        noise) may not be perfectly accurate. HeadPose's pitch value is a
+        reserved field and will always return 0.
+        * Different 'recognitionModel' values are provided. If follow-up
+        operations like Verify, Identify, Find Similar are needed, please
+        specify the recognition model with 'recognitionModel' parameter. The
+        default value for 'recognitionModel' is 'recognition_01', if latest
+        model needed, please explicitly specify the model you need in this
+        parameter. Once specified, the detected faceIds will be associated with
+        the specified recognition model. More details, please refer to [How to
+        specify a recognition
+        model](https://docs.microsoft.com/en-us/azure/cognitive-services/face/face-api-how-to-topics/specify-recognition-model)
+        .
 
         :param url: Publicly reachable URL of an image
         :type url: str
@@ -338,6 +457,19 @@ class FaceOperations(object):
          time cost.
         :type return_face_attributes: list[str or
          ~azure.cognitiveservices.vision.face.models.FaceAttributeType]
+        :param recognition_model: Name of recognition model. Recognition model
+         is used when the face features are extracted and associated with
+         detected faceIds, (Large)FaceList or (Large)PersonGroup. A recognition
+         model name can be provided when performing Face - Detect or
+         (Large)FaceList - Create or (Large)PersonGroup - Create. The default
+         value is 'recognition_01', if latest model needed, please explicitly
+         specify the model you need. Possible values include: 'recognition_01',
+         'recognition_02'
+        :type recognition_model: str or
+         ~azure.cognitiveservices.vision.face.models.RecognitionModel
+        :param return_recognition_model: A value indicating whether the
+         operation should return 'recognitionModel' in response.
+        :type return_recognition_model: bool
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: returns the direct response alongside the
          deserialized response
@@ -366,6 +498,10 @@ class FaceOperations(object):
             query_parameters['returnFaceLandmarks'] = self._serialize.query("return_face_landmarks", return_face_landmarks, 'bool')
         if return_face_attributes is not None:
             query_parameters['returnFaceAttributes'] = self._serialize.query("return_face_attributes", return_face_attributes, '[FaceAttributeType]', div=',')
+        if recognition_model is not None:
+            query_parameters['recognitionModel'] = self._serialize.query("recognition_model", recognition_model, 'str')
+        if return_recognition_model is not None:
+            query_parameters['returnRecognitionModel'] = self._serialize.query("return_recognition_model", return_recognition_model, 'bool')
 
         # Construct headers
         header_parameters = {}
@@ -470,7 +606,7 @@ class FaceOperations(object):
     verify_face_to_person.metadata = {'url': '/verify'}
 
     def detect_with_stream(
-            self, image, return_face_id=True, return_face_landmarks=False, return_face_attributes=None, custom_headers=None, raw=False, callback=None, **operation_config):
+            self, image, return_face_id=True, return_face_landmarks=False, return_face_attributes=None, recognition_model="recognition_01", return_recognition_model=False, custom_headers=None, raw=False, callback=None, **operation_config):
         """Detect human faces in an image and returns face locations, and
         optionally with faceIds, landmarks, and attributes.
 
@@ -490,6 +626,19 @@ class FaceOperations(object):
          time cost.
         :type return_face_attributes: list[str or
          ~azure.cognitiveservices.vision.face.models.FaceAttributeType]
+        :param recognition_model: Name of recognition model. Recognition model
+         is used when the face features are extracted and associated with
+         detected faceIds, (Large)FaceList or (Large)PersonGroup. A recognition
+         model name can be provided when performing Face - Detect or
+         (Large)FaceList - Create or (Large)PersonGroup - Create. The default
+         value is 'recognition_01', if latest model needed, please explicitly
+         specify the model you need. Possible values include: 'recognition_01',
+         'recognition_02'
+        :type recognition_model: str or
+         ~azure.cognitiveservices.vision.face.models.RecognitionModel
+        :param return_recognition_model: A value indicating whether the
+         operation should return 'recognitionModel' in response.
+        :type return_recognition_model: bool
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: returns the direct response alongside the
          deserialized response
@@ -521,6 +670,10 @@ class FaceOperations(object):
             query_parameters['returnFaceLandmarks'] = self._serialize.query("return_face_landmarks", return_face_landmarks, 'bool')
         if return_face_attributes is not None:
             query_parameters['returnFaceAttributes'] = self._serialize.query("return_face_attributes", return_face_attributes, '[FaceAttributeType]', div=',')
+        if recognition_model is not None:
+            query_parameters['recognitionModel'] = self._serialize.query("recognition_model", recognition_model, 'str')
+        if return_recognition_model is not None:
+            query_parameters['returnRecognitionModel'] = self._serialize.query("return_recognition_model", return_recognition_model, 'bool')
 
         # Construct headers
         header_parameters = {}