+++ azure-cognitiveservices-vision-face/HISTORY.rst
@@ -3,6 +3,36 @@
 Release History
 ===============
 
+0.3.0 (2019-03-28)
+++++++++++++++++++
+
+**Features**
+
+- Model PersonGroup has a new parameter recognition_model
+- Model FaceList has a new parameter recognition_model
+- Model LargePersonGroup has a new parameter recognition_model
+- Model LargeFaceList has a new parameter recognition_model
+- Model DetectedFace has a new parameter recognition_model
+
+**Breaking changes**
+
+- Operation FaceListOperations.create has a new signature
+- Operation FaceListOperations.get has a new signature
+- Operation FaceOperations.detect_with_stream has a new signature
+- Operation FaceOperations.detect_with_stream has a new signature
+- Operation FaceOperations.detect_with_url has a new signature
+- Operation FaceOperations.detect_with_url has a new signature
+- Operation LargeFaceListOperations.create has a new signature
+- Operation LargeFaceListOperations.get has a new signature
+- Operation LargePersonGroupOperations.create has a new signature
+- Operation LargePersonGroupOperations.get has a new signature
+- Operation LargePersonGroupOperations.list has a new signature
+- Operation PersonGroupOperations.create has a new signature
+- Operation PersonGroupOperations.get has a new signature
+- Operation PersonGroupOperations.list has a new signature
+- Operation FaceListOperations.list has a new signature
+- Operation LargeFaceListOperations.list has a new signature
+
 0.2.0 (2019-03-07)
 ++++++++++++++++++